requirements.txt
filetype==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==9.5.0
toml==0.10.2
